+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Upcoming release
 
+- `codeframe` formatter now adds final newline in output.
+
 ## 1.1.0 (2019-06-04)
 
 - `input-missing-label` now validates `<textarea>` and `<select>`.